+++ R/make_adapter.R
@@ -36,5 +36,9 @@ make_adapter <- function(spec, data_list){
     return(QuantileAdapter$new(spec, dat))
   }
 
+  if(inherits(spec, "jc_spec_netbenefit")){
+    return(NetBenefitAdapter$new(spec, dat))
+  }
+
   stop("Unsupported spec type: ", paste(class(spec), collapse = ", "))
 }